[bugfix] remove AboutInteractor
