+++ VolleyBolley/Modules/About/DI/AboutAssembly.swift
@@ -12,9 +12,7 @@ final class AboutAssembly: Assembly {
 	func assemble(container: Container) {
 		container.register(AboutViewController.self) { _ in
 			let router = AboutRouter()
-			let interactor = AboutInteractor()
 			let presenter = AboutPresenter(
-				interactor: interactor,
 				router: router
 			)
 			let view = AboutViewController(presenter: presenter)
